Allow passing DirectX 11 textures to OpenVR
This uses the types defined in the `windows` crate. ID3D11Texture2D is a
COM interface. The Interface trait implement an `as_raw()` method. This
is the pointer that needs to be passed to the SteamVR compositor at
submit time.

Gated the functionality behind a `submit_d3d11` feature, but also
enabled it by default for convenience. We can change that.